+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow [Semantic Versions](https://semver.org/).
 
 - Add configuration for django `manage.py` file path
 - Improve `pre-commit.run-hooks` command with `params`
+- Add fallback for `poetry.update` and `poetry.update_to_latest`
 
 ## 0.9.1