+++ CHANGELOG.md
@@ -4,8 +4,13 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## unreleased
 
+- Add configuration for django `manage.py` file path
 - Improve `pre-commit.run-hooks` command with `params`
+<<<<<<< HEAD
 - Add `git.blame-copy` command
+=======
+- Add fallback for `poetry.update` and `poetry.update_to_latest`
+>>>>>>> main
 
 ## 0.9.1